+++ roseau/load_flow/io/dgs/switches.py
@@ -1,8 +1,10 @@
 import logging
+import warnings
 
 import pandas as pd
 import shapely
 
+from roseau.load_flow.exceptions import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
 from roseau.load_flow.models import Bus, Switch
 from roseau.load_flow.typing import Id
 
@@ -30,11 +32,21 @@ def generate_switches(
     has_geometry = "GPSlon" in elm_coup.columns and "GPSlat" in elm_coup.columns
     for switch_id in elm_coup.index:
         # TODO: use the detailed phase information instead of n
-        n = elm_coup.at[switch_id, "nphase"] + elm_coup.at[switch_id, "nneutral"]
-        phases = "abc" if n == 3 else "abcn"
+        nphase = elm_coup.at[switch_id, "nphase"]
+        nneutral = elm_coup.at[switch_id, "nneutral"]
+        if nphase != 3:
+            msg = f"nphase={nphase!s} for switch {switch_id!r} is not supported. Only 3-phase switches are currently supported."
+            logger.error(msg)
+            raise RoseauLoadFlowException(msg=msg, code=RoseauLoadFlowExceptionCode.DGS_BAD_PHASE_NUMBER)
+        phases = "abcn" if nneutral else "abc"
         bus1 = buses[sta_cubic.at[elm_coup.at[switch_id, "bus1"], "cterm"]]
         bus2 = buses[sta_cubic.at[elm_coup.at[switch_id, "bus2"], "cterm"]]
         geometry = (
             shapely.Point(elm_coup.at[switch_id, "GPSlon"], elm_coup.at[switch_id, "GPSlat"]) if has_geometry else None
         )
+        on_off = elm_coup.at[switch_id, "on_off"]
+        if not on_off:
+            warnings.warn(
+                f"Switch {switch_id!r} is open but switches are always closed in roseau-load-flow.", stacklevel=4
+            )
         switches[switch_id] = Switch(id=switch_id, phases=phases, bus1=bus1, bus2=bus2, geometry=geometry)